+++ src/main/java/conspire/Conspire.java
@@ -33,6 +33,7 @@ import conspire.cards.colorless.SpireCoStock;
 import conspire.cards.curse.Blindness;
 import conspire.cards.green.DoublingDagger;
 import conspire.cards.green.PoisonWeapons;
+import conspire.cards.red.Charge;
 import conspire.cards.red.ExplosiveBarrier;
 import conspire.cards.red.HitWhereItHurts;
 import conspire.cards.red.Purge;
@@ -50,6 +51,7 @@ import conspire.potions.EchoDraught;
 import conspire.potions.TimeTravelPotion;
 import conspire.relics.Boomerang;
 import conspire.relics.DecoderRing;
+import conspire.relics.Dentures;
 import conspire.relics.Flyswatter;
 import conspire.relics.GiftBox;
 import conspire.relics.GlowingRock;
@@ -138,6 +140,7 @@ public class Conspire implements
     @Override
     public void receiveEditCards() {
         // red
+        BaseMod.addCard(new Charge());
         BaseMod.addCard(new ExplosiveBarrier());
         BaseMod.addCard(new HitWhereItHurts());
         BaseMod.addCard(new Purge());
@@ -161,6 +164,7 @@ public class Conspire implements
     public void receiveEditRelics() {
         BaseMod.addRelic(new Boomerang(), RelicType.SHARED);
         BaseMod.addRelic(new DecoderRing(), RelicType.SHARED);
+        BaseMod.addRelic(new Dentures(), RelicType.SHARED);
         BaseMod.addRelic(new Flyswatter(), RelicType.SHARED);
         BaseMod.addRelic(new GiftBox(), RelicType.SHARED);
         BaseMod.addRelic(new GlowingRock(), RelicType.SHARED);